+++ R/tree_accessors.R
@@ -1960,14 +1960,14 @@ setMethod(
 #' @export
 setGeneric(
   "coltree",
-  function(obj, df = NULL, rtpos = TreePos()) standardGeneric("coltree")
+  function(obj, df = NULL, rtpos = TreePos(), alt_counts_df = df) standardGeneric("coltree")
 )
 
 #' @rdname col_accessors
 #' @exportMethod coltree
 setMethod(
   "coltree", "InstantiatedColumnInfo",
-  function(obj, df = NULL, rtpos = TreePos()) {
+  function(obj, df = NULL, rtpos = TreePos(), alt_counts_df = df) {
     if (!is.null(df)) {
       warning("Ignoring df argument and retrieving already-computed LayoutColTree")
     }
@@ -1979,14 +1979,14 @@ setMethod(
 #' @export coltree
 setMethod(
   "coltree", "PreDataTableLayouts",
-  function(obj, df, rtpos) coltree(clayout(obj), df, rtpos)
+  function(obj, df, rtpos, alt_counts_df = df) coltree(clayout(obj), df, rtpos, alt_counts_df = alt_counts_df)
 )
 
 #' @rdname col_accessors
 #' @export coltree
 setMethod(
   "coltree", "PreDataColLayout",
-  function(obj, df, rtpos) {
+  function(obj, df, rtpos, alt_counts_df = df) {
     obj <- set_def_child_ord(obj, df)
     kids <- lapply(
       obj,
@@ -1994,7 +1994,8 @@ setMethod(
         splitvec_to_coltree(
           df = df,
           splvec = x,
-          pos = rtpos
+          pos = rtpos,
+          alt_counts_df = alt_counts_df
         )
       }
     )
@@ -2005,7 +2006,8 @@ setMethod(
         lev = 0L,
         kids = kids,
         tpos = rtpos,
-        spl = RootSplit()
+        spl = RootSplit(),
+        colcount = NROW(alt_counts_df)
       )
     }
     disp_ccounts(res) <- disp_ccounts(obj)
@@ -2017,21 +2019,21 @@ setMethod(
 #' @export coltree
 setMethod(
   "coltree", "LayoutColTree",
-  function(obj, df, rtpos) obj
+  function(obj, df, rtpos, alt_counts_df) obj
 )
 
 #' @rdname col_accessors
 #' @export coltree
 setMethod(
   "coltree", "VTableTree",
-  function(obj, df, rtpos) coltree(col_info(obj))
+  function(obj, df, rtpos, alt_counts_df) coltree(col_info(obj))
 )
 
 #' @rdname col_accessors
 #' @export coltree
 setMethod(
   "coltree", "TableRow",
-  function(obj, df, rtpos) coltree(col_info(obj))
+  function(obj, df, rtpos, alt_counts_df) coltree(col_info(obj))
 )
 
 setGeneric("coltree<-", function(obj, value) standardGeneric("coltree<-"))
@@ -2148,7 +2150,17 @@ setGeneric("col_counts", function(obj, path = NULL) standardGeneric("col_counts"
 #' @rdname col_accessors
 setMethod(
   "col_counts", "InstantiatedColumnInfo",
-  function(obj, path = NULL) obj@counts[.path_to_pos(path, obj, cols = TRUE)]
+  function(obj, path = NULL) {
+    if (is.null(path)) {
+      lfs <- collect_leaves(coltree(obj))
+      ret <- vapply(lfs, facet_colcount, 1L, path = NULL)
+    } else {
+      ret <- facet_colcount(obj, path)
+    }
+    ## required for strict backwards compatibility,
+    ## even though its undesirable behavior.
+    unname(ret)
+  }
 )
 
 #' @export
@@ -2167,7 +2179,25 @@ setGeneric("col_counts<-", function(obj, path = NULL, value) standardGeneric("co
 setMethod(
   "col_counts<-", "InstantiatedColumnInfo",
   function(obj, path = NULL, value) {
-    obj@counts[.path_to_pos(path, obj, cols = TRUE)] <- value
+    ## obj@counts[.path_to_pos(path, obj, cols = TRUE)] <- value
+    ## obj
+    if (!is.null(path)) {
+      all_paths <- list(path)
+    } else {
+      all_paths <- make_col_df(obj, visible_only = TRUE)$path
+    }
+    if (length(value) != length(all_paths)) {
+      stop("Got ", length(value), "values for ",
+        length(all_paths), " column paths",
+        if (is.null(path)) " (from path = NULL)",
+        "."
+      )
+    }
+    ctree <- coltree(obj)
+    for (i in seq_along(all_paths)) {
+      facet_colcount(ctree, all_paths[[i]]) <- value[i]
+    }
+    coltree(obj) <- ctree
     obj
   }
 )
@@ -2256,6 +2286,24 @@ setMethod(
 )
 
 #' @rdname int_methods
+setMethod(
+  "disp_ccounts", "LayoutColTree",
+  function(obj) obj@display_columncounts
+)
+
+#' @rdname int_methods
+setMethod(
+  "disp_ccounts", "LayoutColLeaf",
+  function(obj) obj@display_columncounts
+)
+
+#' @rdname int_methods
+setMethod(
+  "disp_ccounts", "Split",
+  function(obj) obj@child_show_colcounts
+)
+
+#' @rdname int_methods
 setGeneric("disp_ccounts<-", function(obj, value) standardGeneric("disp_ccounts<-"))
 
 #' @rdname int_methods
@@ -2298,11 +2346,319 @@ setMethod(
 
 #' @rdname int_methods
 setMethod(
+  "disp_ccounts<-", "LayoutColLeaf",
+  function(obj, value) {
+    obj@display_columncounts <- value
+    obj
+  }
+)
+
+#' @rdname int_methods
+setMethod(
   "disp_ccounts<-", "PreDataTableLayouts",
   function(obj, value) {
     clyt <- clayout(obj)
     disp_ccounts(clyt) <- value
     clayout(obj) <- clyt
+    obj
+  }
+)
+
+match_path_by_pos <- function(kidlst, path) {
+  ret <- -1
+  nmval_pairs <- lapply(
+    kidlst,
+    function(kd) {
+      pos <- tree_pos(kd)
+      c(obj_name(tail(pos_splits(pos), 1)[[1]]),
+        rawvalues(tail(pos_splvals(pos), 1))[[1]])
+    }
+  )
+
+  matches <- vapply(
+    nmval_pairs,
+    function(pair) {
+      (pair[1] == path[1]) && (path[2] %in% c(pair[2], "*"))
+    },
+    TRUE
+  )
+  if (any(matches))
+    ret <- which(matches)
+  ret
+
+}
+
+## close to a duplicate of tt_at_path, but... not quite :(
+#' @rdname int_methods
+coltree_at_path <- function(obj, path, ...) {
+  if (length(path) == 0)
+    return(obj)
+  stopifnot(
+    is(path, "character"),
+    length(path) > 0,
+    !anyNA(path)
+  )
+  if (any(grepl("@content", path, fixed = TRUE)))
+    stop("@content token is not valid for column paths.")
+
+
+  ## if(obj_name(obj) == path[1]) {
+  ##   path <- path[-1]
+  ## }
+
+  cur <- obj
+  curpath <- path
+  while (length(curpath) > 0) {
+    kids <- tree_children(cur)
+    kidmatch <- match_path_by_pos(kids, curpath)
+    curname <- curpath[1]
+    if (curname == "@content")
+      stop("@content is not a valid token for a column path")
+    else if (kidmatch > 0)
+      cur <- kids[[kidmatch]]
+    else
+      stop("Path appears invalid for this tree at step ", curname)
+    curpath <- curpath[-(1:2)]  # name and value, both consumed due to structure
+  }
+  cur
+}
+
+`coltree_at_path<-` <- function(obj, path, value) {
+  obj <- coltree(obj) ## noop if it already is
+  if (any(grepl("@content", path, fixed = TRUE)))
+    stop("@content token is not valid for column paths.")
+  ## we don't have intermediary structures in the coltree model
+  ## **after the first one** Yes this is bad and wrong but
+
+  ## its how it is.
+  ## ie it goes straight from A: Drug X -> F
+  ## in the split_cols_by("ARM") %>% split_cols_by("SEX") case
+  ## but we want to use the do_recursive_replace machinery
+  ## that is already well tested
+  trimmed_path <- path[c(1, seq(2, length(path), by = 2))]
+
+  do_recursive_replace(obj, trimmed_path, value = value)
+}
+
+
+#' Set visibility of column counts for a group of sibling facets
+#'
+#' @inheritParams gen_args
+#' @param path character. The path *to the parent of the
+#' desired siblings*. The last element in the path should
+#' be a split name.
+#' @return obj, modified with the desired columncount
+#' display behavior
+`facet_colcounts_visible<-` <- function(obj, path, value) {
+  coldf <- make_col_df(obj, visible_only = FALSE)
+  allpaths <- coldf$path
+  lenpath <- length(path)
+  match_paths <- vapply(allpaths, function(path_i) {
+    (length(path_i) == lenpath + 1) &&
+      (all(head(path_i, -1) == path))
+  }, TRUE)
+  for (curpath in allpaths[match_paths]) {
+    colcount_visible(obj, curpath) <- value
+  }
+  obj
+}
+
+#' Get or set column count for a facet in column space
+#'
+#' @inheritParams gen_args
+#' @param path character. This path must end on a
+#' split value, e.g., the level of a categorical variable
+#' that was split on in column space, but it need not
+#' be the path to an individual column.
+#'
+#' @return for `facet_colcount` the current count associated
+#' with that facet in column space, for `facet_colcount<-`,
+#' `obj` modified with the new column count for the specified
+#' facet.
+#'
+#' @note Updating a lower-level (more specific)
+#' column count manually **will not** update the
+#' counts for its parent facets. This cannot be made
+#' automatic because the rtables framework does not
+#' require sibling facets to be mutually exclusive
+#' (e.g., total "arm", faceting into cumulative
+#' quantiles, etc) and thus the count of a parent facet
+#' will not always be simply the sum of the counts for
+#' all of its children.
+#'
+#' @export
+#' @examples
+#' lyt <- basic_table() %>%
+#'   split_cols_by("ARM", show_colcounts = TRUE) %>%
+#'   split_cols_by("SEX", split_fun = keep_split_levels(c("F", "M")),
+#'     show_colcounts = TRUE) %>%
+#'   split_cols_by("STRATA1", show_colcounts = TRUE) %>%
+#'   analyze("AGE")
+#'
+#' tbl <- build_table(lyt, ex_adsl)
+#'
+#' facet_colcount(tbl, c("ARM", "A: Drug X"))
+#' facet_colcount(tbl, c("ARM", "A: Drug X", "SEX", "F"))
+#' facet_colcount(tbl, c("ARM", "A: Drug X", "SEX", "F", "STRATA1", "A"))
+#'
+#' ## modify specific count after table creation
+#' facet_colcount(tbl, c("ARM", "A: Drug X", "SEX", "F", "STRATA1", "A")) <- 25
+#'
+#' ## show black space for certain counts by assign NA
+#'
+#' facet_colcount(tbl, c("ARM", "A: Drug X", "SEX", "F", "STRATA1", "C")) <- NA
+setGeneric("facet_colcount",
+           function(obj, path) standardGeneric("facet_colcount"))
+
+#' @rdname facet_colcount
+#' @export
+setMethod("facet_colcount", "LayoutColTree",
+  function(obj, path = NULL) {
+    ## if(length(path) == 0L)
+    ##   stop("face_colcount requires a non-null path") #nocov
+    subtree <- coltree_at_path(obj, path)
+    subtree@column_count
+  }
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod("facet_colcount", "LayoutColLeaf",
+  function(obj, path = NULL) {
+    ## not sure if we should check for null here as above
+    obj@column_count
+  }
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount", "VTableTree",
+  function(obj, path) facet_colcount(coltree(obj), path = path)
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount", "InstantiatedColumnInfo",
+  function(obj, path) facet_colcount(coltree(obj), path = path)
+)
+
+#' @rdname facet_colcount
+#' @export
+setGeneric(
+  "facet_colcount<-",
+  function(obj, path, value) standardGeneric("facet_colcount<-")
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount<-", "LayoutColTree",
+  function(obj, path, value) {
+    ct <- coltree_at_path(obj, path)
+    ct@column_count <- as.integer(value)
+    coltree_at_path(obj, path) <- ct
+    obj
+  }
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount<-", "LayoutColLeaf",
+  function(obj, path, value) {
+    obj@column_count <- as.integer(value)
+    obj
+  }
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount<-", "VTableTree",
+  function(obj, path, value) {
+    cinfo <- col_info(obj)
+    facet_colcount(cinfo, path) <- value
+    col_info(obj) <- cinfo
+    obj
+  }
+)
+
+#' @rdname facet_colcount
+#' @export
+setMethod(
+  "facet_colcount<-", "InstantiatedColumnInfo",
+  function(obj, path, value) {
+    ct <- coltree(obj)
+    facet_colcount(ct, path) <- value
+    coltree(obj) <- ct
+    obj
+  }
+)
+
+#' Value and Visibility of specific column counts by path
+#'
+#' @inheritParams gen_args
+#'
+#' @return for `colcount_visible` a logical scalar
+#' indicating whether the specified position in
+#' the column hierarchy is set to display its column count;
+#' for `colcount_visible<-`, `obj` updated with
+#' the specified count displaying behavior set.
+#'
+#' @note Users generally should not call `colcount_visible`
+#' directly, as setting sibling facets to have differing
+#' column count visibility will result in an error when
+#' printing or paginating the table.
+#'
+#' @export
+setGeneric("colcount_visible", function(obj, path) standardGeneric("colcount_visible"))
+
+#' @rdname colcount_visible
+#' @export
+setMethod("colcount_visible", "VTableTree",
+          function(obj, path) colcount_visible(coltree(obj), path))
+
+#' @rdname colcount_visible
+#' @export
+setMethod("colcount_visible", "InstantiatedColumnInfo",
+          function(obj, path) colcount_visible(coltree(obj), path))
+
+#' @rdname colcount_visible
+#' @export
+setMethod(
+  "colcount_visible", "LayoutColTree",
+  function(obj, path) {
+    subtree <- coltree_at_path(obj, path)
+    disp_ccounts(subtree)
+  }
+)
+
+#' @rdname colcount_visible
+#' @export
+setGeneric("colcount_visible<-", function(obj, path, value) standardGeneric("colcount_visible<-"))
+
+#' @rdname colcount_visible
+#' @export
+setMethod(
+  "colcount_visible<-", "VTableTree",
+  function(obj, path, value) {
+    ctree <- coltree(obj)
+    colcount_visible(ctree, path) <- value
+    coltree(obj) <- ctree
+    obj
+  }
+)
+
+#' @rdname colcount_visible
+#' @export
+setMethod(
+  "colcount_visible<-", "LayoutColTree",
+  function(obj, path, value) {
+    subtree <- coltree_at_path(obj, path)
+    disp_ccounts(subtree) <- value
+    coltree_at_path(obj, path) <- subtree
     obj
   }
 )
@@ -2338,6 +2694,14 @@ setMethod(
   "colcount_format", "PreDataTableLayouts",
   function(obj) colcount_format(clayout(obj))
 )
+
+#' @rdname int_methods
+#' @export
+setMethod(
+  "colcount_format", "Split",
+  function(obj) obj@child_colcount_format
+)
+
 
 #' @rdname int_methods
 #' @export
@@ -2386,6 +2750,56 @@ setMethod(
     clyt <- clayout(obj)
     colcount_format(clyt) <- value
     clayout(obj) <- clyt
+    obj
+  }
+)
+
+## It'd probably be better if this had the full set of methods as above
+## but its not currently modelled in the class and probably isn't needed
+## super much
+#' @rdname int_methods
+#' @export
+setGeneric("colcount_na_str", function(obj) standardGeneric("colcount_na_str"))
+
+#' @rdname int_methods
+#' @export
+setMethod(
+  "colcount_na_str", "InstantiatedColumnInfo",
+  function(obj) obj@columncount_na_str
+)
+
+#' @rdname int_methods
+#' @export
+setMethod(
+  "colcount_na_str", "VTableNodeInfo",
+  function(obj) colcount_na_str(col_info(obj))
+)
+
+#' @rdname int_methods
+#' @export
+setGeneric(
+  "colcount_na_str<-",
+  function(obj, value) standardGeneric("colcount_na_str<-")
+)
+
+#' @export
+#' @rdname int_methods
+setMethod(
+  "colcount_na_str<-", "InstantiatedColumnInfo",
+  function(obj, value) {
+    obj@columncount_na_str <- value
+    obj
+  }
+)
+
+#' @rdname int_methods
+#' @export
+setMethod(
+  "colcount_na_str<-", "VTableNodeInfo",
+  function(obj, value) {
+    cinfo <- col_info(obj)
+    colcount_na_str(cinfo) <- value
+    col_info(obj) <- cinfo
     obj
   }
 )